% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_theme.R
\name{apply_theme}
\alias{apply_theme}
\title{Apply IOM theme to a plot}
\usage{
apply_theme(type, basesize = 10, facets = FALSE)
}
\arguments{
\item{type}{What type of plot is being themed? Accepts "line",
"bar-horizontal", "bar-vertical", "map", and "void".}

\item{basesize}{Font size in points. Default is 10.}

\item{facets}{Is this a faceted plot?}
}
\description{
Adds thematic presets adhering to the IOM visual branding to a ggplot object.
}
\details{
The \code{basesize} parameter is used for the font size of axis labels and titles
and plot annotations.
}
