% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_lines.R
\name{break_lines}
\alias{break_lines}
\title{Add line breaks to string}
\usage{
break_lines(text, maxchars, strip = NULL)
}
\arguments{
\item{text}{String to add line breaks to.}

\item{maxchars}{Maximum number of characters per line.}
}
\description{
Add line breaks to string
}
