% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check available data for country}
\usage{
check_data(country)
}
\arguments{
\item{country}{3-letter ISO code of country of interest.}
}
\value{
Tibble.
}
\description{
Display the variables and years for which data is available for a given
country.
}
