% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryname.R
\name{countryname}
\alias{countryname}
\title{Toggle between country names and ISO codes}
\usage{
countryname(key, from = "iso3c", to = "unname")
}
\arguments{
\item{key}{Name or ISO code of country.}

\item{from}{Current type of \code{key}.}

\item{to}{Target type of \code{key}.}
}
\value{
String or vector of strings.
}
\description{
See \code{countrynames} table for full dictionary.
}
\examples{
countryname("DEU")
# Germany

countryname("Germany", from = "unname", to = "iso3c")
# DEU

}
