% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disrupt.R
\docType{data}
\name{disrupt}
\alias{disrupt}
\title{Biophysical disruptions dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60735 rows and 5 columns.
}
\usage{
disrupt
}
\description{
\code{R} dataset measuring the impact of various environmental disasters.
Categories are people affected, people displaced, people killed, and economic
damage.
}
\keyword{datasets}
