% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_source.R
\name{format_source}
\alias{format_source}
\title{Format plot source in HTML}
\usage{
format_source(source, basesize = 8, space_after = TRUE, color = "#8099D0")
}
\arguments{
\item{source}{String of the plot's source.}

\item{basesize}{Base text size of plot.}

\item{space_after}{Whether to add line breaks after the source. Use when plot has a caption.}

\item{color}{Text color.}
}
\value{
String.
}
\description{
Format plot source in HTML
}
