% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi_plot_data.R
\name{gdi_plot_data}
\alias{gdi_plot_data}
\title{Grab data behind a preset plot}
\usage{
gdi_plot_data(key, iso)
}
\arguments{
\item{key}{Plot key.}

\item{iso}{Country.}
}
\description{
Grab data behind a preset plot
}
