% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdidata.R
\docType{data}
\name{gdidata}
\alias{gdidata}
\title{List of datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 5 columns.
}
\usage{
gdidata
}
\description{
All available datasets.
}
\keyword{datasets}
