% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdiplots.R
\docType{data}
\name{gdiplots}
\alias{gdiplots}
\title{List of charts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
gdiplots
}
\description{
All charts that can be generated by \code{gdiplot()}.
}
\keyword{datasets}
