% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdiviz-package.R
\docType{package}
\name{gdiviz-package}
\alias{gdiviz}
\alias{gdiviz-package}
\title{gdiviz: Tools for IOM-themed Plots}
\description{
Tools for generating aesthetic plots with the International Organization for Migration visual branding. Developed and maintained by the IOM Global Data Insitute.
}
\keyword{internal}
