% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{Get members of a country group}
\usage{
get_members(group)
}
\arguments{
\item{group}{(Case-insensitive) name of grouping.}
}
\value{
Vector of ISO codes.
}
\description{
Currently accepts "EU", "Schengen", "Balkans".
}
