% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_geo.R
\docType{data}
\name{idmc_geo}
\alias{idmc_geo}
\title{Geolocated nternal displacements dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26857 rows and 7 columns.
}
\source{
Internal Displacement Monitoring Centre
(https://www.internal-displacement.org/database/displacement-data/)
}
\usage{
idmc_geo
}
\description{
Flow count of internally displaced persons with coordinates, by cause of
displacement.
}
\keyword{datasets}
