% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linemarker.R
\name{linemarker}
\alias{linemarker}
\title{Insert line marker in plot}
\usage{
linemarker(direction, at = 0)
}
\arguments{
\item{direction}{"h" for a horizontal line or "v" for a vertical line.}

\item{at}{x- or y-intercept.}
}
\description{
Insert line marker in plot
}
