% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmp.R
\docType{data}
\name{mmp}
\alias{mmp}
\title{Missing migrants dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16765 rows and 9 columns.
}
\source{
International Organization for Migration Missing Migrants Project
(https://missingmigrants.iom.int/downloads)
}
\usage{
mmp
}
\description{
Dataset of migrants who die or go missing during the act of international
migration, by migrant origin, location of incident, and cause of death. It
draws from official records, media reports, non-government organizations, and
interviews.
}
\keyword{datasets}
