% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpp.R
\docType{data}
\name{mpp}
\alias{mpp}
\title{Missing migrants dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16765 rows and 9 columns.
}
\usage{
mpp
}
\description{
\code{R} dataset containing the Missing Migrants Project dataset, which tracks
migrants who die or go missing during the act of international migration.
}
\keyword{datasets}
