% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal}
\alias{pal}
\title{Choose an IOM color}
\usage{
pal(color, shade = 1)
}
\arguments{
\item{color}{Accepts "blues", "unblues", "yellows", "greens", "oranges",
"reds", and "grays".}

\item{shade}{The shade of the chosen color, from 1 (darkest) to 5 (lightest).
Accepts vectors.}
}
\value{
Hex code/s as a string or vector of strings.
}
\description{
Choose an IOM color
}
\examples{

pal("blues")
pal("reds", c(1, 3, 5))

}
