% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizer.R
\name{sizer}
\alias{sizer}
\title{Get size parameters for plot}
\usage{
sizer(basesize)
}
\arguments{
\item{basesize}{Font size in pt.}
}
\value{
List comprising \code{text} for general text, \code{title} for plot title,
\code{stext} for slightly smaller text, and \code{footnote} for caption text.
}
\description{
Get size parameters for plot
}
