% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Migrant stocks dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 255927 rows and 5 columns.
}
\usage{
stocks
}
\description{
\code{R} dataset containing stocks of migrants by country-pairs at
5-year intervals from 1990 to 2020, disaggregated by sex. Run \code{gdidata} for
more information.
}
\keyword{datasets}
