% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp.R
\docType{data}
\name{wpp}
\alias{wpp}
\title{Population by sex and age dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 637200 rows and 5 columns.
}
\usage{
wpp
}
\description{
Population count by sex and age from the UN World Population Prospects 2022
update. Data is at 10-year intervals from 1950 to 2020, plus 2021.
}
\keyword{datasets}
