% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrynames.R
\docType{data}
\name{countrynames}
\alias{countrynames}
\title{List of countries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 242 rows and 7 columns.
}
\usage{
countrynames
}
\description{
\describe{
\item{iso3n, iso2c, iso3c}{ISO country codes}
\item{eurostat, wb}{Other country codes}
\item{unname}{Officially recognized name}
\item{unname_sort}{Name used for sorting}
\item{popname}{Popular name}
\item{region}{World Bank region}
}
}
\keyword{datasets}
