% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi_plot.R
\name{gdi_plot}
\alias{gdi_plot}
\title{Generate a preset plot}
\usage{
gdi_plot(plot_key, iso, basesize = 7, title, dims = FALSE)
}
\arguments{
\item{plot_key}{Plot key.}

\item{iso}{Country.}

\item{basesize}{Base text size.}

\item{title}{Plot title.}

\item{dims}{If \code{TRUE}, returns the default plot dimensions in cm
(default\verb{=FALSE}).}
}
\value{
\code{ggplot2} object.
}
\description{
Generate a preset plot
}
\examples{
gdi_plot("stocks", "DEU")

}
