% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdidata.R
\docType{data}
\name{gdidata}
\alias{gdidata}
\title{List of datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 7 columns.
}
\usage{
gdidata
}
\description{
This lists all the datasets bundled with \code{gdiviz}. \code{indicators} groups
together all single-variable time series.
}
\keyword{datasets}
