% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_plot_data.R
\name{grab_plot_data}
\alias{grab_plot_data}
\title{Grab data behind a preset plot}
\usage{
grab_plot_data(plot_key, iso, years = FALSE, year_max = FALSE, print = FALSE)
}
\arguments{
\item{plot_key}{Plot key.}

\item{iso}{Country.}

\item{years}{If \code{TRUE}, returns a string showing the year range of the data
(default\verb{=FALSE}).}

\item{year_max}{If \code{TRUE}, returns the maximum year of the data
(default\verb{=FALSE}).}

\item{print}{If \code{TRUE}, returns the dataset formatted for printing.}
}
\description{
Grab data behind a preset plot
}
\examples{
grab_plot_data("stocks", "DEU")

}
