% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{Dataset of country indicators}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 76614 rows and 4 columns.
}
\usage{
indicators
}
\description{
Dataset of country-level time series of the following variables:
\itemize{
\item Birth rate
\item Age dependency ratio
\item FDI inflow
\item FDI outflow
\item Group Grievance Index
\item GDP per capita
\item Inflation rate
\item Net migration
\item Political Stability and Absence of Violence/Terrorism
\item Population
\item Remittances received
\item Remittances paid
\item Unemployment rate
}

Run \code{gdidata} for more information.
}
\keyword{datasets}
